feat: Subindo modulo do consumo do Pub/Sub com db

+++ backend-v1/src/main/java/br/com/fatec/messaging/model/OrderItem.java
@@ -1,7 +1,5 @@
 package br.com.fatec.messaging.model;
 
-import java.math.BigDecimal;
-
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -14,16 +12,17 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "order_item")
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class OrderItem {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @ManyToOne
     @JoinColumn(name = "order_uuid", nullable = false)
